Server: Add unique constraint on name and owner ID of items table
In theory, the server enforces uniquness because when doing a PUT
operation it either creates the items if it doesn't exist, or overwrite
it. However, there's race condition that makes it possible for multiple
items with the same name being created per user. So we add this
constraint to ensure that any additional query would fail (which can be
recovered by repeating the request).