DO NOT create screenshots of text !!!
If there's text to select, do it. Why create a screenshot? Screenshots are helpful for GUI/UI/UX elements, but useless and highly annoting for text. Is copying text really that hard and complicated?

Oce upon a time I asked a customer to send me an error message (so that I could search for it in the code base and in traces). That customer created a screenshot of a terminal, with only half the error message and then put that image into a word document and sent that to me. Besides the fact that I did not have Word, what would I have done with an image? Argh!!!!!!!